requirements.txt
uvicorn[standard]
python-dotenv
websockets



